% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_graph.R
\name{knn_graph}
\alias{knn_graph}
\title{Returns the adjacency matrix of a knn graph from the given distance matrix}
\usage{
knn_graph(data, k)
}
\arguments{
\item{data}{a matrix of points as rows}

\item{k}{number of nearest neighbors to compute}
}
\description{
Returns the adjacency matrix of a knn graph from the given distance matrix
}
