% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization_score.R
\name{localization_score}
\alias{localization_score}
\title{Ranks features by how localized they are in graph}
\usage{
localization_score(adj_matrix, f, num_perms = 1000, num_cores = 1,
  perm_estimate = F)
}
\arguments{
\item{adj_matrix}{a (preferrably sparse) binary matrix of adjacency between the columns of f}

\item{f}{a numeric vector or matrix specifying one or more functions with support on
the set of points whose significance will be assesed in the simplicial complex. Each
column corresponds to a point and each row specifies a different function.}

\item{num_perms}{number of permutations used to build the null distribution for each
feature. By default is set to 1000.}

\item{num_cores}{integer specifying the number of cores to be used in the computation. By
default only one core is used.}

\item{perm_estimate}{boolean indicating whether normal distribution parameters should be
determined from num_perms permutations to estimate the p-value. By default is set to FALSE.}
}
\description{
Given the adjacency matrix of a graph and a set of features on that graph, ranks those features (f)
by the equation f((e^{kA}-I)/k)g, which measures how much those features are colocalized in the graph.
Calculates the p-value of this score by permuting the columns of the feature matrix.
}
