% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_score.R
\name{overlap_score}
\alias{overlap_score}
\title{Ranks pairs of features using the Combinatorial Laplacian Score for 0- and 1-forms.}
\usage{
overlap_score(f, f_pairs, num_perms = 1000, seed = 10, num_cores = 1)
}
\arguments{
\item{f}{a numeric vector or matrix specifying one or more functions with support on
the set of points whose significance will be assesed in the simplicial complex. Each
column corresponds to a point and each row specifies a different function.}

\item{f_pairs}{a 2 column matrix where each row specifes the indices or names
of a pair of points on which the Comb. Lap. score will be computed}

\item{num_perms}{number of permutations used to build the null distribution for each
feature. By default is set to 1000.}

\item{seed}{integer specifying the seed used to initialize the generator of permutations.
By default is set to 10.}

\item{num_cores}{integer specifying the number of cores to be used in the computation. By
default only one core is used.}
}
\description{
Given a nerve or a clique complex, a set of features consisting of functions with support on
the set of points underlying the complex, and a list of pairs of features,
it asseses the significance of each pair of features
in the simplicial complex by computing its scalar and vectorial Combinatorial Laplacian
Score and comparing it with the null distribution that results from reshufling many times the values of
the function across the point cloud. For nerve complexes, feature functions induce 0- and
1-forms in the complex by averaging the function across the points associated to 0- and 1-simplices
respectively. For clique complexes, feature functions are directly 0-forms in the complex and 1-forms
are obtained by averaging the function across the two vertices connected by each edge.
}
