% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_score2.R
\name{adjacency_score2}
\alias{adjacency_score2}
\title{Ranks pairs of features by how localized they are in graph}
\usage{
adjacency_score2(adj_matrix, f, f_pairs, c, num_perms = 1000, num_cores = 1,
  perm_estimate = F, groupings = F)
}
\arguments{
\item{adj_matrix}{a (preferrably sparse) binary matrix of adjacency between the columns of f}

\item{f}{a numeric vector or matrix specifying one or more functions with support on
the set of points whose significance will be assesed in the simplicial complex. Each
column corresponds to a point and each row specifies a different function.}

\item{f_pairs}{a 2 column matrix where each row specifes the indices or names
of a pair of points on which the Comb. Lap. score will be computed}

\item{c}{constant used to determine width of diffusion, must be 0 <= c}

\item{num_perms}{number of permutations used to build the null distribution for each
feature. By default is set to 1000.}

\item{num_cores}{integer specifying the number of cores to be used in the computation. By
default only one core is used.}

\item{perm_estimate}{boolean indicating whether normal distribution parameters should be
determined from num_perms permutations to estimate the p-value. By default is set to FALSE.}

\item{groupings}{boolean indicating whether features are binary and mutually exclusive
indicated each point's inclusion in some group. Allows for p-value computation from a
parameterized hypergeometric null distribution. By default is set to FALSE.}
}
\description{
Given the adjacency matrix of a graph and a set of features on that graph, ranks given pairs
of those features (f and g) by the equation f((e^{kA}-I)/k)g, which measures how much those
features are colocalized in the graph. Calculates the p-value for this score by permuting
the columns of the feature matrix separately for each features.
}
